FractalWindow refactoring, phase 2
Problem: Finally, it's impossible for 'FractalWindow' to inherit from
'QWidget' and 'FractalWindowOGL' to inherit from 'QGLWidget' and
'FractalWindow' at the same time. For that, 'QGLWidget' must have a
virtual inheritance from 'QWidget', and this is not the case.

Solution: Thus, 'FractalWindow' inherit from 'QGLWidget' directly, and
'FractalWindowOGL' inherit also of 'FractalWindow'. But, since
'FractalWindow' is designed to be an interface for OpenGL and Cairo
classes, it's an abstract class and haven't the behavior of a
'QGLWidget', but the one of a 'QWidget'.