Documentation: makefile warnings update
Problem:
Warnings are shown when we create documentation from root directory. But
if we call 'make doc' from root directory, it's not to correct it or
debug it, but just to watch it.

Fix:
Show warnings only if we generate documentation from 'doc' subdirectory.
Otherwise, we just generate documentation normally.