bug correction:
We can now draw a new fractal if we delete the previous one we the key escape.
It works when the previous fractal is in a tab or in
another window (with key F2 or F3).
It also works when we close the window (F3 mode) of the previous fractal
with the close button.

Other bug:
	- same bug with the close button of tabs
	- draw more than one fractal at a time